Remove ramp parameters sorting
Maya sorts the ramp positions, colors, not by ramp order but by the
order in which you create these elements. Sorting these globally though,
would break some ramp cases, such as the box ramp, so this needs to be
disabled, and the component access is dealt with in the shader code.