Add new IDs, node help, bump UI fixes
The bump control automagically allows the connection of a bump2d, but it
inherits this from the common section of the Maya's Lambert node and
doesn't allow split bump controls. Adding a new procedure we can now
have different bump for different layers, i.e, coating, substrate.
The plugin was not parsing standard OSL URL metadata token, and as a
result, the default node help was launching the autodesk page and trying
to find our node's documentation there. Adding this allows the user to
go direct from maya attribute editor to the appleseed read-the-docs
documentation.
Finally, new IDs added for the Substance PBR material and the Fabric
material (closure WIP).